// src/modules/indicators/infraestructure/persistence/models/orders.model.ts
export class OrdersModel {
  PorcentajeEmbarcado: number;
  PorcentajeFacturado: number;
  OrdenCompra: string;
  Mr: string;
  Pedido: string;
  OrgVenta: string;
  Estatus: string;
  FechaOc: string;
  Nocte: string;
  Cliente: string;
  NocteDestino: string;
  ClienteDestino: string;
  Zona: string;
  RegionCte: string;
  UOperativa: string;
  CantidadPedido: number;
  MontoLocal: number;
  CantidadFacturada: number;
  MontoFacturado: number;
  AttachementCount: number;
}